+++ autoclasswrapper/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.1.12"
+__version__ = "0.1.13"
 __copyright__ = "Copyright 2018 Pierre Poulain"
 
 import logging