+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 - print Exception content for check_data_type()
 - output class probability for every gene/protein
 - simplify the calculation of cluster stats
-
+- class/cluster numbering starts at 1 (0-based in autoclass output)
 
 **0.1.12**
 - update write_cluster_stats() methods with mean/std column